Refactor tag command
